third_party/blink/web_tests/platform/mac/external/wpt/css/css-grid/grid-extrinsically-sized-mutations-expected.txt

This is a testharness.js-based test.
[FAIL] grid-extrinsically-sized-mutations
  assert_equals: height expected 40 but got 44
[FAIL] grid-extrinsically-sized-mutations 4
  assert_equals: height expected 30 but got 22
[FAIL] grid-extrinsically-sized-mutations 6
  assert_equals: height expected 30 but got 33
[FAIL] grid-extrinsically-sized-mutations 7
  assert_equals: height expected 40 but got 22
Harness: the test ran to completion.
